refactor: Skill 상수 클래스 생성자 private로 변경
