+++ src/main/java/sixgaezzang/sidepeek/skill/exception/message/SkillErrorMessage.java
@@ -1,7 +1,11 @@
-package sixgaezzang.sidepeek.skill.util.validation;
+package sixgaezzang.sidepeek.skill.exception.message;
 
 import static sixgaezzang.sidepeek.projects.util.ProjectConstant.MAX_CATEGORY_LENGTH;
 
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class SkillErrorMessage {
     public static final String SKILL_NOT_EXISTING = "Skill Id에 해당하는 스킬이 없습니다.";
     public static final String SKILL_IS_NULL = "기술 스택이 null 입니다.";
