+++ src/main/java/sixgaezzang/sidepeek/skill/util/validation/SkillValidator.java
@@ -3,9 +3,9 @@ package sixgaezzang.sidepeek.skill.util.validation;
 import static sixgaezzang.sidepeek.common.util.ValidationUtils.validateMaxLength;
 import static sixgaezzang.sidepeek.common.util.ValidationUtils.validateNotBlank;
 import static sixgaezzang.sidepeek.projects.util.ProjectConstant.MAX_CATEGORY_LENGTH;
-import static sixgaezzang.sidepeek.skill.util.validation.SkillErrorMessage.CATEGORY_IS_NULL;
-import static sixgaezzang.sidepeek.skill.util.validation.SkillErrorMessage.CATEGORY_OVER_MAX_LENGTH;
-import static sixgaezzang.sidepeek.skill.util.validation.SkillErrorMessage.SKILL_IS_NULL;
+import static sixgaezzang.sidepeek.skill.exception.message.SkillErrorMessage.CATEGORY_IS_NULL;
+import static sixgaezzang.sidepeek.skill.exception.message.SkillErrorMessage.CATEGORY_OVER_MAX_LENGTH;
+import static sixgaezzang.sidepeek.skill.exception.message.SkillErrorMessage.SKILL_IS_NULL;
 
 import io.jsonwebtoken.lang.Assert;
 import sixgaezzang.sidepeek.skill.domain.Skill;
